Compatibilidad de respaldos

+++ src/accesoDatos/UpdateVersion.java
@@ -11,6 +11,7 @@ import interfase.menus.Menu;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
+import logica.utilitarios.Ut;
 
 /**
  *
@@ -55,7 +56,10 @@ public class UpdateVersion {
         } // end if (!UtilBD.fieldInTable(conn, "enviarFacturaE", "config"))
 
         // Bosco agregado 18/07/2019.  Elimino varios índices y agrego uno.
-        if (UtilBD.indexInDB(conn, "FK_Hinmovimd_bodexis")) {
+        double versionNumber = 
+                Double.parseDouble(Ut.quitarCaracteres(Menu.dataBaseVersion, ".").toString());
+        
+        if (versionNumber >= 5.7 && UtilBD.indexInDB(conn, "FK_Hinmovimd_bodexis")) {
             sqlSent
                     = "ALTER TABLE `inmovimd`  "
                     + "        DROP INDEX `FK_Hinmovimd_bodexis`";
@@ -65,7 +69,7 @@ public class UpdateVersion {
             ps.close();
         } // end if
 
-        if (UtilBD.indexInDB(conn, "Index_Movdocu_H")) {
+        if (versionNumber >= 5.7 && UtilBD.indexInDB(conn, "Index_Movdocu_H")) {
             sqlSent
                     = "ALTER TABLE `inmovimd`  "
                     + "        DROP INDEX `Index_Movdocu_H`";
@@ -75,7 +79,7 @@ public class UpdateVersion {
             ps.close();
         } // end if
 
-        if (UtilBD.indexInDB(conn, "FK_Hinmovimd_inmovime")) {
+        if (versionNumber >= 5.7 && UtilBD.indexInDB(conn, "FK_Hinmovimd_inmovime")) {
             sqlSent
                     = "ALTER TABLE `inmovimd`  "
                     + "        DROP INDEX `FK_Hinmovimd_inmovime`";
@@ -85,7 +89,7 @@ public class UpdateVersion {
             ps.close();
         } // end if
         
-        if (UtilBD.indexInDB(conn, "FK_Hinmovim_Tipocambio")) {
+        if (versionNumber >= 5.7 && UtilBD.indexInDB(conn, "FK_Hinmovim_Tipocambio")) {
             sqlSent
                     = "ALTER TABLE `inmovime`  "
                     + "        DROP INDEX `FK_Hinmovim_Tipocambio`";
@@ -95,7 +99,7 @@ public class UpdateVersion {
             ps.close();
         } // end if
         
-        if (!UtilBD.indexInDB(conn, "Index_recalcular_inv")) {
+        if (versionNumber >= 5.7 && !UtilBD.indexInDB(conn, "Index_recalcular_inv")) {
             sqlSent
                     = "ALTER TABLE `inmovime`  "
                     + "        ADD INDEX `Index_recalcular_inv` (`movfech` ASC, `estado` ASC)";
@@ -105,7 +109,7 @@ public class UpdateVersion {
             ps.close();
         } // end if
         
-        if (UtilBD.indexInDB(conn, "fk_hbodexis_hintarticu")) {
+        if (versionNumber >= 5.7 && UtilBD.indexInDB(conn, "fk_hbodexis_hintarticu")) {
             sqlSent
                     = "ALTER TABLE `hbodexis`  "
                     + "        DROP FOREIGN KEY `fk_hbodexis_hintarticu`";
@@ -115,7 +119,7 @@ public class UpdateVersion {
             ps.close();
         } // end if
         
-        if (UtilBD.indexInDB(conn, "FK_hbodexis_bodexis")) {
+        if (versionNumber >= 5.7 && UtilBD.indexInDB(conn, "FK_hbodexis_bodexis")) {
             sqlSent
                     = "ALTER TABLE `hbodexis`  "
                     + "        DROP INDEX `FK_hbodexis_bodexis`";
